python/bitplane/pyproject.toml
[project]
name = "bitplane"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
	"matplotlib",
	"numpy",
	"pillow",
]

[project.scripts]
bitplane = "bitplane:main"